requirements.txt
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openai>=1.40.0
httpx<0.28.0
PyPDF2==3.0.1
pymupdf>=1.24.0
# Optional for OCR fallback on image-only PDFs: pip install pytesseract Pillow; install Tesseract (e.g. brew install tesseract)
python-dotenv==1.0.0
watchdog==3.0.0
tiktoken>=0.5.0
anthropic>=0.39.0
openpyxl>=3.1.0
google-genai>=1.0.0
google-generativeai>=0.8.0
python-docx>=1.0.0
requests>=2.28.0